Update patch set 1
Patch Set 1:

Mecha and Ace have been able to properly use OpenGL for all applications except Browser, which needed to be individually modified for CPU Upload with a toggle switch. The issues with GMail and other "typically" effected applications did not occur when OpenGL was properly enabled in Webkit. It still comes down to the same point that rendering is only temporarily broken, but by using an ifdef for QCOM_HARDWARE you prevent any future disabling of this individual hack without reverting the patch altogether. Much like MISSING_EGL_EXTERNAL_IMAGE, which became obsolete through new driver, is it better to have to alter code and submit a new patch or just remove a flag?

Patch-set: 1