Update patch set 1
Patch Set 1:

you still did not answer the question... and from the answer u did give. basically u said... there is not a qcom device, that is using the gl renderer, that does not need the patch, and you are speculating that there will be a need for this after the issue is resolved. which would lead to another patch like this one. what if the solution is to nuke all this together and a implementation of a real fix?

Patch-set: 1