Update patch set 1
Patch Set 1:

I answered your question in the first sentence. It seems you are not trying to discuss the validity of the patch, but simply arguing that a patch you cherry-picked is sufficient. Again, would you rather simply remove a board flag and maintain code that a future device may still make use of or submit patches that will simply need to be reverted later? I am modifying the code for CM9 efficiency, not to simply follow in the footsteps of an alternate coding group.

Patch-set: 1