Update patch set 1
Patch Set 1:

no u +1 the patch i submited only to revert it... whats the sense in that? why not just submit this in the first place? why did u submit basically the same patch to begin with, and now say its wrong in some way? its all over the a board config option.... i dont see this being useful in any way after its officially fixed. i just submitted a patch that fixes the issue. -1 it if u dont like it :)

Patch-set: 1