Update patch set 2
Patch Set 2:

Given the comment on the code

"The default / alternative engine depends on the device class. On devices with a lot of memory (e.g. Passion/Sholes), the default is V8. On everything else, the only choice is JSC.", and sholes having 256M of RAM

and

"need a flag to tell the C side when we're on devices with large memory budgets (i.e. larger than the low-end devices that initially shipped)"

and this commit citing >= 256M http://paste.ubuntu.com/792377/

I thought it'd be good for buzz/click/zeppelin/morrison/motus. Particularly I've been running this on zeppelin and it seems to be a bit faster when browsing and rendering/scrolling webpages. I didn't do any real measuring with numbers though.

And yeah, I failed big time when mentioning tegra, and that's why there's a patchset 2 :p But I still think the patch has a use case.

Patch-set: 2