Update patch set 3
Patch Set 3:

Hrm... V8 originally needed armv7, but we fixed that a long time ago. My armv6 devices, including a 7225, are happily using v8, you only need to add JS_ENGINE to the board config with no additional flag.

As for the memory thing, it's only relevant for image quality decisions (color and format conversions; the most critical change is that it moves the file limit in which to do runtime from 2MB to 8MB); it should probably be tested to check if it's worth it, since the displays on the devices we're talking about aren't that good to begin with...

Patch-set: 3